requirements.txt
asgiref==3.7.2
boto3==1.28.25
botocore==1.31.25
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.10.0
django-froala-editor==4.1.1
django-storages==1.13.2
jmespath==1.0.1
Pillow==10.0.0
python-dateutil==2.8.2
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
urllib3==1.26.16
django-jazzmin==2.5.0